pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "inbox_zero"
version = "0.1.0"
description = "Inbox Zero"
authors = [{ name = "Kishen Sharma", email = "ksharma06@pm.me" }]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "flask>=3.1",
    "pydantic>=2.11",
    "openai>=2.0",
    "openrouter>=0.8",
    "langchain>=1.2",
    "beautifulsoup4>=4.14",
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.3",
    "redis>=7.0",
    "slack_bolt>=1.28",
    "tiktoken>=0.12",
    "python-dotenv",
    "PyYAML>=6.0",
    "uuid_utils",
]

[dependency-groups]
dev = [
    "pytest-mock>=3.15",
    "pytest-cov>=5.0",
    "pip-audit>=2.7",
    "ruff>=0.4",
    "pre-commit>=4.5",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
known-third-party = ["src"]
